class structure reorganization
