+++ src/main/java/App.java
@@ -1,5 +1,5 @@
-import org.mitre.synthea.modules.CommunityHealthWorker;
-import org.mitre.synthea.modules.Generator;
+import org.mitre.synthea.engine.Generator;
+import org.mitre.synthea.world.agents.CommunityHealthWorker;
 
 /*
  * This Java source file was generated by the Gradle 'init' task.
