+++ src/main/java/org/mitre/synthea/engine/EventList.java
@@ -1,4 +1,4 @@
-package org.mitre.synthea.modules;
+package org.mitre.synthea.engine;
 
 import java.util.ArrayList;
 import java.util.List;
