+++ src/main/java/org/mitre/synthea/modules/Immunizations.java
@@ -10,7 +10,10 @@ import java.util.Map;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
-import org.mitre.synthea.modules.HealthRecord.Entry;
+import org.mitre.synthea.helpers.Utilities;
+import org.mitre.synthea.world.agents.Person;
+import org.mitre.synthea.world.concepts.HealthRecord;
+import org.mitre.synthea.world.concepts.HealthRecord.Entry;
 
 import com.google.gson.Gson;
